Split convergence.py into focused modules
convergence.py had grown to roughly 3180 lines, the largest module in
the package, and the new JSON solve and memo cache formed a
self-contained seam worth isolating. Extract the cache into
_convergence_cache.py, which depends only on the standard library and
numpy, and the per-iteration mesh building and refinement into
_convergence_meshing.py, which carries the geometry and movement
dependencies. convergence.py keeps the two public analyze functions and
their small metric and validation helpers, so the public API is
unchanged.

Following the _fixed_point_relaxation precedent, the functions that
form each private module's interface take bare names while the
module-internal helpers keep their leading underscores. Split the unit
tests so each module's tests live in a file named for it, and update
the architecture registry in CLAUDE.md. This is a pure code move with
no behavior change, and the full test suite passes unchanged.
